+++ NEWS.md
@@ -1,6 +1,6 @@
 # targets 1.3.2.9000 (development)
 
-* Extra safeguard in the `crew` algorithm to opt back into the default RNG kind.
+* Add function `tar_seed_set()` which sets a seed and the default RNG algorithms.
 
 # targets 1.3.2